requirements.txt
fastapi
uvicorn
streamlit
google-genai
python-dotenv
tavily-python
PyPDF2
python-docx
python-multipart
pytest
pytest-asyncio
httpx
pypdf
chromadb
sentence-transformers
